Fix OnRemoveConnection from calling side
If we use the original:
if (port != null && port.IsConnectedTo(this)) port.node.OnRemoveConnection(port);
we going to have always wrong condition, about "IsConnectedTo(this)" and port.node.OnRemoveConnection(port) just won't be called, cause we already have disconnected ports. So we need to do it exactly at the moment of disconnection.